+++ core/src/main/java/com/graphhopper/routing/ev/DefaultImportRegistry.java
@@ -20,9 +20,7 @@ package com.graphhopper.routing.ev;
 
 import com.graphhopper.routing.util.*;
 import com.graphhopper.routing.util.parsers.*;
-
-import java.util.Arrays;
-import java.util.stream.Collectors;
+import com.graphhopper.util.PMap;
 
 public class DefaultImportRegistry implements ImportRegistry {
     @Override
@@ -99,6 +97,11 @@ public class DefaultImportRegistry implements ImportRegistry {
             return ImportUnit.create(name, props -> MaxLength.create(),
                     (lookup, props) -> new OSMMaxLengthParser(
                             lookup.getDecimalEncodedValue(MaxLength.KEY))
+            );
+        else if (Orientation.KEY.equals(name))
+            return ImportUnit.create(name, props -> Orientation.create(),
+                    (lookup, props) -> new OrientationCalculator(
+                            lookup.getDecimalEncodedValue(Orientation.KEY))
             );
         else if (Surface.KEY.equals(name))
             return ImportUnit.create(name, props -> Surface.create(),
@@ -209,15 +212,17 @@ public class DefaultImportRegistry implements ImportRegistry {
 
         else if (BusAccess.KEY.equals(name))
             return ImportUnit.create(name, props -> BusAccess.create(),
-                    (lookup, props) -> new ModeAccessParser(TransportationMode.BUS, lookup.getBooleanEncodedValue(BusAccess.KEY),
-                            lookup.getBooleanEncodedValue(Roundabout.KEY), Arrays.stream(props.getString("restrictions", "").split(";")).filter(s -> !s.isEmpty()).collect(Collectors.toList())),
+                    (lookup, props) -> new ModeAccessParser(TransportationMode.BUS,
+                            lookup.getBooleanEncodedValue(name), true, lookup.getBooleanEncodedValue(Roundabout.KEY),
+                            PMap.toSet(props.getString("restrictions", "")), PMap.toSet(props.getString("barriers", ""))),
                     "roundabout"
             );
 
         else if (HovAccess.KEY.equals(name))
             return ImportUnit.create(name, props -> HovAccess.create(),
-                    (lookup, props) -> new ModeAccessParser(TransportationMode.HOV, lookup.getBooleanEncodedValue(HovAccess.KEY),
-                            lookup.getBooleanEncodedValue(Roundabout.KEY), Arrays.stream(props.getString("restrictions", "").split(";")).filter(s -> !s.isEmpty()).collect(Collectors.toList())),
+                    (lookup, props) -> new ModeAccessParser(TransportationMode.HOV,
+                            lookup.getBooleanEncodedValue(name), true, lookup.getBooleanEncodedValue(Roundabout.KEY),
+                            PMap.toSet(props.getString("restrictions", "")), PMap.toSet(props.getString("barriers", ""))),
                     "roundabout"
             );
         else if (FootTemporalAccess.KEY.equals(name))