+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-### 9.0 [not yet released]
+### 10.0 [not yet released]
+
+- The config-example.yml uses a non-empty snap_preventions default array: [tunnel, bridge and ferry] for the /route endpoint
+- the default u-turn time is now 0, the default u-turn weight is still infinite 
+- turn restriction support for restrictions with overlapping and/or multiple via-edges/ways, #3030
+- constructor of BaseGraph.Builder uses byte instead of integer count.
+- KeyValue is now KValue as it holds the value only. Note, the two parameter constructor uses one value for the forward and one for the backward direction (and no longer "key, value")
+- sac_scale priority handling for bicycles moved to the bike custom models
+
+### 9.0 [23 Apr 2024]
 
 - max_slope is now a signed decimal, see #2955
 - move sac_scale handling out of foot_access parser and made foot safer via lowering to sac_scale<2, same for hike sac_scale<5